separate service and cli
